+++ src/main/java/com/new_bank_app/services/AccountService.java
@@ -15,7 +15,7 @@ public class AccountService {
         this.accountRepository = accountRepository;
     }
 
-    public Account getAccount(int accountId) {
+    public Account getAccount(Long accountId) {
         return accountRepository.getAccountById(accountId);
     }
