+++ src/main/java/com/new_bank_app/services/ValidationService.java
@@ -16,12 +16,12 @@ public class ValidationService {
         this.accountRepository = accountRepository;
     }
 
-    public boolean checkWithdrawGreaterThanBalance(int accountId, BigDecimal withdrawAmountValue, User user) {
+    public boolean checkWithdrawGreaterThanBalance(Long accountId, BigDecimal withdrawAmountValue, User user) {
         BigDecimal currentBalance = accountRepository.getAccountBalance(user.getUserId(), accountId);
         return (currentBalance.compareTo(withdrawAmountValue) < 0);
     }
 
-    public boolean checkTransferGreaterThanBalance(int transferFromId, BigDecimal transferAmount,  User user) {
+    public boolean checkTransferGreaterThanBalance(Long transferFromId, BigDecimal transferAmount,  User user) {
         BigDecimal currentBalanceAccountFrom = accountRepository.getAccountBalance(user.getUserId(), transferFromId);
         return (currentBalanceAccountFrom.compareTo(transferAmount) < 0);
     }
